+++ src/PatpassComercio/Exceptions/InscriptionStartException.php
