+++ src/PatpassComercio/Exceptions/InscriptionStatusException.php
