+++ src/PatpassComercio/Responses/InscriptionStatusResponse.php
