+++ src/Webpay/TransaccionCompleta/Responses/TransactionCaptureResponse.php
