+++ src/Webpay/TransaccionCompleta/Responses/TransactionDetail.php
